refactor: Listener 관련 코드를 listener 디렉토리로 이동
