+++ module-api/src/main/java/com/trybe/moduleapi/challenge/event/listener/ChallengePreferenceEventListener.java
@@ -1,5 +1,7 @@
-package com.trybe.moduleapi.challenge.event;
+package com.trybe.moduleapi.challenge.event.listener;
 
+import com.trybe.moduleapi.challenge.event.ChallengeEvent;
+import com.trybe.moduleapi.challenge.event.ChallengeEventType;
 import com.trybe.modulecore.challenge.entity.Challenge;
 import com.trybe.modulecore.challenge.repository.preference.ChallengePreferenceCache;
 import org.springframework.scheduling.annotation.Async;
